Update README.md
This project aims to predict the prices of used cars based on various features such as the year of manufacture, the current ex-showroom price, kilometers driven, and more. The predictions are made using different machine learning models, and the performance of these models is evaluated to determine the best approach.